Add combat minigame
