+++ myning/chapters/mine/actions/combat.py
@@ -1,0 +1,255 @@
+import random
+from dataclasses import dataclass
+from functools import lru_cache
+
+from rich.table import Table
+
+from myning.chapters.mine.actions import Action
+from myning.chapters.mine.combat_minigame import CombatMinigame
+from myning.objects.army import Army
+from myning.objects.character import Character
+from myning.objects.player import Player
+from myning.objects.settings import Settings
+from myning.objects.stats import IntegerStatKeys, Stats
+from myning.objects.trip import Trip
+from myning.utilities.file_manager import FileManager
+from myning.utilities.formatter import Formatter
+from myning.utilities.generators import generate_enemy_army, generate_reward
+from myning.utilities.tab_title import TabTitle
+from myning.utilities.ui import Icons
+
+player = Player()
+settings = Settings()
+stats = Stats()
+trip = Trip()
+
+
+class CombatAction(Action):
+    # pylint: disable=redefined-builtin
+    def __init__(self, *, enemies: Army | None = None, round=1):
+        if enemies:
+            self.enemies = enemies
+        else:
+            assert trip.mine
+            if trip.mine.enemies[1] < 0:
+                trip.mine.enemies[1] = len(player.army) + trip.mine.enemies[1]
+            self.enemies = generate_enemy_army(
+                trip.mine.character_levels,
+                trip.mine.enemies,
+                trip.mine.max_enemy_items,
+                trip.mine.max_enemy_item_level,
+                trip.mine.enemy_item_scale,
+            )
+        self.round = round
+        self.round_logs: list[RoundLog] = []
+        self.damage_done = 0
+        self.damage_taken = 0
+        TabTitle.change_tab_subactivity(
+            f"⚔️ Battling ({player.species.icon}{len(player.army.living_members)} "
+            f"v 👽{len(self.enemies.living_members)})"
+        )
+        self.duration = random.randint(5, 9)
+        self.game = CombatMinigame(self)
+        super().__init__(self.duration)
+
+    @property
+    def content(self):
+        if not settings.mini_games_disabled:
+            return self.game
+        allies = Army(player.army.living_members)
+        enemies = Army(self.enemies.living_members)
+        table = Table.grid()
+        table.add_row("[orange1]Oh no! You're under attack[/]\n")
+        table.add_row(f"[bold]Round {self.round}[/]")
+        table.add_row(f"Fighting... ({self.duration} seconds left)\n")
+        table.add_row("⚔️   " * (5 - (self.duration - 1) % 5))
+        table.add_row("\n[bold]Your Army[/]")
+        table.add_row(allies.compact_view if settings.compact_mode else allies.battle_view)
+        table.add_row("\n[bold]Enemy Army[/]")
+        table.add_row(enemies.compact_view if settings.compact_mode else enemies.battle_view)
+        return table
+
+    def fight(self, bonus: float):
+        battle_order = _get_battle_order(player.army.living_members, self.enemies.living_members)
+        # bonus = _mini_game_bonus(static_menu)
+        for attacker in battle_order:
+            if player.army.defeated or self.enemies.defeated:
+                break
+
+            is_friendly = attacker in player.army
+            defender = random.choice(
+                self.enemies.living_members if is_friendly else player.army.living_members
+            )
+            if is_friendly:
+                damage, dodged, crit = _calculate_damage(attacker, defender, bonus)
+            else:
+                damage, dodged, crit = _calculate_damage(attacker, defender)
+
+            if damage > 0:
+                if not dodged:
+                    defender.subtract_health(damage)
+                self.round_logs.append(
+                    RoundLog(
+                        attacker=attacker,
+                        defender=defender,
+                        is_friendly=is_friendly,
+                        damage=damage,
+                        dodged=dodged,
+                        critted=crit,
+                    )
+                )
+            if is_friendly:
+                self.damage_done += damage
+                FileManager.save(attacker)
+            else:
+                self.damage_taken += damage
+                FileManager.save(defender)
+
+            if defender.health <= 0:
+                battle_order.remove(defender)
+                if is_friendly:
+                    stats.increment_int_stat(IntegerStatKeys.FALLEN_SOLDIERS)
+                    FileManager.save(stats)
+
+    @property
+    @lru_cache(maxsize=1)
+    def next(self):
+        self.fight(self.game.bonus)
+        self.game.remove()
+        if player.army.defeated:
+            return None
+        if self.enemies.defeated:
+            trip.add_battle(len(self.enemies), True)
+            FileManager.save(trip)
+            return VictoryAction(len(self.enemies))
+        next_combat_action = CombatAction(enemies=self.enemies, round=self.round + 1)
+        return RoundAction(
+            self.game.bonus,
+            self.damage_done,
+            self.damage_taken,
+            self.round_logs,
+            next_combat_action,
+        )
+
+
+@dataclass
+class RoundLog:
+    attacker: Character
+    defender: Character
+    is_friendly: bool
+    damage: int
+    dodged: bool
+    critted: bool
+
+
+class RoundAction(Action):
+    def __init__(
+        self,
+        bonus: float,
+        damage_done: int,
+        health_lost: int,
+        round_logs: list[RoundLog],
+        next_action: Action,
+    ):
+        damage_width = max(len(str(damage_done)), len(str(health_lost)))
+        self.summary = "[bold]Round Summary[/]\n\n"
+        if not settings.mini_games_disabled:
+            self.summary += f"{_bonus_str(bonus)}\n\n"
+        self.summary += (
+            f"[bold green1]{damage_done:{damage_width}}[/] damage inflicted.\n"
+            f"[bold red1]{health_lost:{damage_width}}[/] damage sustained.\n"
+        )
+        self.log_table = Table.grid(padding=(0, 1, 0, 0))
+        if round_logs:
+            for log in round_logs:
+                self.log_table.add_row(
+                    str(log.attacker.icon),
+                    f"[{'green1' if log.is_friendly else 'red1'}]{log.attacker.name}[/]",
+                    Icons.CRIT if log.critted else "",
+                    "[bold dark_cyan](0)[/]"
+                    if log.dodged
+                    else f"[{'bold orange1' if log.critted else 'normal'}]{log.damage}[/]",
+                    Icons.DODGE if log.dodged else "",
+                    str(log.defender.icon),
+                    f"[{'red1' if log.is_friendly else 'green1'}]{log.defender.name}[/]",
+                    Icons.DEATH if log.defender.health <= 0 else "",
+                )
+            self.log_table.columns[3].justify = "right"
+        self.next_action = next_action
+        super().__init__(5)
+
+    @property
+    def content(self):
+        table = Table.grid(expand=True)
+        table.add_row(self.summary)
+        table.add_row(f"Returning to battle in {self.duration}...\n")
+        table.add_row(self.log_table)
+        return table
+
+    @property
+    def next(self):
+        return self.next_action
+
+
+class VictoryAction(Action):
+    def __init__(self, enemy_count: int):
+        TabTitle.change_tab_subactivity("")
+        assert trip.mine
+        rewards = generate_reward(trip.mine.max_item_level, enemy_count)
+        trip.add_items(*rewards)
+        FileManager.multi_save(trip, *rewards)
+        self.rewards = rewards
+        super().__init__(len(rewards) + 1)
+
+    @property
+    def content(self):
+        lines = ["[green1]You won the battle![/]\n"] + [
+            item.battle_new_str for item in self.rewards[: len(self.rewards) - self.duration + 1]
+        ]
+        return "\n".join(lines)
+
+    @property
+    def next(self):
+        return self if self.duration > 1 else None
+
+
+def _get_battle_order(allies: list[Character], enemies: list[Character]):
+    combined = allies + enemies
+    random.shuffle(combined)
+    return combined
+
+
+def _calculate_damage(attacker: Character, defender: Character, bonus: float = 1):
+    dodge_chance = int(defender.stats["dodge_chance"])
+    dodge = random.choices(
+        [True, False],
+        weights=[dodge_chance, 100 - dodge_chance],
+    )[0]
+
+    critical_chance = int(attacker.stats["critical_chance"])
+    crit = random.choices(
+        [True, False],
+        weights=[critical_chance, 100 - critical_chance],
+    )[0]
+
+    damage = attacker.stats["damage"]
+    damage = random.randint(0, damage)
+    damage = int(bonus * damage)
+    if crit:
+        damage *= 2
+
+    armor = defender.stats["armor"]
+    blocked = random.randint(0, max(armor, 0))
+    damage -= blocked
+    damage = max(damage, 0)
+    return damage, dodge, crit
+
+
+def _bonus_str(bonus: float):
+    return (
+        "[bold]No[/] damage bonus"
+        if bonus == 1
+        else f"[bold green1]{Formatter.percentage(bonus)}[/] damage bonus"
+        if bonus > 1
+        else f"[bold red1]{Formatter.percentage(bonus)}[/] damage penalty"
+    )
